Move `useGetAvailability` to it's own file
