+++ src/components/availability-label/helper.ts
@@ -3,8 +3,8 @@ import clsx from "clsx";
 import { useGetV1ProductsIdentifier } from "../../core/publizon/publizon";
 import { useConfig } from "../../core/utils/config";
 import { AccessTypeCode } from "../../core/dbc-gateway/generated/graphql";
-import { useGetAvailability } from "../../apps/material/helper";
 import { FaustId } from "../../core/utils/types/ids";
+import useGetAvailability from "../../core/utils/useGetAvailability";
 
 export const useAvailabilityData = ({
   accessTypes,
